Add text_content_detection tests
